requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
nltk>=3.8.1
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.12.0
jinja2>=3.1.2
python-multipart>=0.0.6
